+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.1.0 (in development)
 - "basic" `tag2version` method: If `regex` is given and it does not contain a
   group named "`version`," the entire text matched by the regex will be used as
   the version
+- "basic" `write` method: `encoding` now defaults to UTF-8
 
 v0.1.0a1 (2021-07-05)
 ---------------------